VTKU-51 : Siirrä metodi mutatoitavaan olioon
Ei toiminnallisia muutoksia, vielä.
