+++ sijoittelu-algoritmi-batch/src/main/java/fi/vm/sade/sijoittelu/batch/logic/impl/algorithm/wrappers/SijoitteluajoWrapper.java
@@ -1,5 +1,10 @@
 package fi.vm.sade.sijoittelu.batch.logic.impl.algorithm.wrappers;
 
+import static fi.vm.sade.sijoittelu.domain.ValintatuloksenTila.EHDOLLISESTI_VASTAANOTTANUT;
+import static fi.vm.sade.sijoittelu.domain.ValintatuloksenTila.KESKEN;
+import static fi.vm.sade.sijoittelu.domain.ValintatuloksenTila.VASTAANOTTANUT_SITOVASTI;
+import static java.util.Arrays.asList;
+import static java.util.Collections.emptyMap;
 import static java.util.Optional.ofNullable;
 import com.google.common.hash.HashCode;
 import com.google.common.hash.HashFunction;
@@ -8,10 +13,19 @@ import com.google.common.hash.Hashing;
 
 import fi.vm.sade.sijoittelu.batch.logic.impl.algorithm.SijoitteluConfiguration;
 import fi.vm.sade.sijoittelu.batch.logic.impl.algorithm.hakukohteet.LisapaikkaTapa;
+import fi.vm.sade.sijoittelu.batch.logic.impl.algorithm.util.TilaTaulukot;
+import fi.vm.sade.sijoittelu.batch.logic.impl.algorithm.util.TilojenMuokkaus;
+import fi.vm.sade.sijoittelu.domain.HakemuksenTila;
+import fi.vm.sade.sijoittelu.domain.Hakemus;
 import fi.vm.sade.sijoittelu.domain.Hakukohde;
 import fi.vm.sade.sijoittelu.domain.SijoitteluAjo;
+import fi.vm.sade.sijoittelu.domain.TilaHistoria;
+import fi.vm.sade.sijoittelu.domain.TilanKuvaukset;
 import fi.vm.sade.sijoittelu.domain.Valintatapajono;
+import fi.vm.sade.sijoittelu.domain.ValintatuloksenTila;
 import fi.vm.sade.sijoittelu.domain.Valintatulos;
+import fi.vm.sade.sijoittelu.domain.dto.VastaanottoDTO;
+import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -19,8 +33,10 @@ import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Consumer;
@@ -264,5 +280,210 @@ public class SijoitteluajoWrapper {
 
     public boolean getHakutoiveidenPriorisointi() {
         return hakutoiveidenPriorisointi;
+    }
+
+    public void paivitaVastaanottojenVaikutusHakemustenTiloihin(Map<String, VastaanottoDTO> aiemmanVastaanotonHakukohdePerHakija, Map<String, Map<String, Map<String, Valintatulos>>> indeksoidutTulokset) {
+        getHakukohteet().forEach(hakukohdeWrapper -> {
+            Map<String, Map<String, Valintatulos>> jonoIndex = indeksoidutTulokset.getOrDefault(hakukohdeWrapper.getHakukohde().getOid(), emptyMap());
+            hakukohdeWrapper.getValintatapajonot().forEach(valintatapajonoWrapper -> {
+                Map<String, Valintatulos> hakemusIndex = jonoIndex.getOrDefault(valintatapajonoWrapper.getValintatapajono().getOid(), emptyMap());
+                valintatapajonoWrapper.getHakemukset().forEach(hakemusWrapper -> {
+                    setHakemuksenValintatuloksenTila(
+                        hakukohdeWrapper,
+                        hakemusWrapper,
+                        hakemusIndex.get(hakemusWrapper.getHakemus().getHakemusOid()),
+                        ofNullable(aiemmanVastaanotonHakukohdePerHakija.get(hakemusWrapper.getHenkilo().getHakijaOid()))
+                    );
+                });
+            });
+        });
+    }
+
+    private static boolean vastaanotonTilaSaaMuuttaaHakemuksenTilaa(Hakemus hakemus) {
+        return TilaTaulukot.kuuluuVastaanotonMuokattavissaTiloihin(hakemus.getEdellinenTila());
+    }
+
+    private static void setHakemuksenValintatuloksenTila(HakukohdeWrapper hakukohdeWrapper,
+                                                         HakemusWrapper hakemusWrapper,
+                                                         Valintatulos valintatulos,
+                                                         Optional<VastaanottoDTO> aiempiVastaanottoSamalleKaudelle) {
+        Hakemus hakemus = hakemusWrapper.getHakemus();
+        LOG.debug("setHakemuksenValintatuloksenTila alkaa: Hakukohde: {}, valintatapajono: {}, hakemus: {}, hakemuksen tila: {}, hakemuksen edellinen tila: {}, vastaanoton tila: {}, aiempi vo: {}",
+                hakemusWrapper.getValintatapajono().getHakukohdeWrapper().getHakukohde().getOid(),
+                hakemusWrapper.getValintatapajono().getValintatapajono().getOid(),
+                hakemus.getHakemusOid(),
+                hakemus.getTila(),
+                hakemus.getEdellinenTila(),
+                valintatulos != null ? valintatulos.getTila() : "(null valintatulos)",
+                aiempiVastaanottoSamalleKaudelle);
+        if (estaaVastaanotonYhdenPaikanSaannoksenTakia(aiempiVastaanottoSamalleKaudelle, hakemusWrapper)) {
+            if (valintatulos != null && valintatulos.getTila() == ValintatuloksenTila.PERUNUT) {
+                TilojenMuokkaus.asetaTilaksiPerunut(hakemusWrapper);
+            } else if (hakemus.getTila() == HakemuksenTila.VARALLA) {
+                hakemus.setTila(HakemuksenTila.PERUUNTUNUT);
+                if (hakemus.getEdellinenTila() != HakemuksenTila.PERUUNTUNUT || hakemus.getTilanKuvaukset() == TilanKuvaukset.tyhja) {
+                    TilojenMuokkaus.asetaTilaksiPeruuntunutVastaanottanutToisenPaikanYhdenPaikanSaannonPiirissa(hakemusWrapper);
+                }
+            }
+            hakemusWrapper.setTilaVoidaanVaihtaa(false);
+        } else if (valintatulos != null && valintatulos.getTila() != null) {
+            if (!vastaanotonTilaSaaMuuttaaHakemuksenTilaa(hakemus)) {
+                // Don't write a log entry
+                valintatulos.setTila(ValintatuloksenTila.KESKEN, ValintatuloksenTila.KESKEN, "", "");
+            }
+            LOG.debug("Hakukohde: {}, valintatapajono: {}, hakemus: {}, hakemuksen tila: {}, hakemuksen edellinen tila: {}, vastaanoton tila: {}",
+                    hakemusWrapper.getValintatapajono().getHakukohdeWrapper().getHakukohde().getOid(),
+                    hakemusWrapper.getValintatapajono().getValintatapajono().getOid(),
+                    hakemus.getHakemusOid(),
+                    hakemus.getTila(),
+                    hakemus.getEdellinenTila(),
+                    valintatulos.getTila());
+            ValintatuloksenTila tila = valintatulos.getTila();
+            boolean voidaanVaihtaa = false;
+            if (tila == ValintatuloksenTila.PERUNUT) {
+                TilojenMuokkaus.asetaTilaksiPerunut(hakemusWrapper);
+            } else if (asList(VASTAANOTTANUT_SITOVASTI, EHDOLLISESTI_VASTAANOTTANUT).contains(tila)) {
+                if (viimeisinHyvaksyttyJono(hakemusWrapper)) {
+                    if (hakemus.getEdellinenTila() == HakemuksenTila.VARALLA || hakemus.getEdellinenTila() == HakemuksenTila.VARASIJALTA_HYVAKSYTTY) {
+                        hyvaksyVarasijalta(hakemusWrapper, valintatulos);
+                    } else {
+                        hyvaksy(hakemusWrapper, valintatulos);
+                    }
+                } else {
+                    valintatulos.setTila(KESKEN, "Peitetään vastaanottotieto ei-hyväksytyltä jonolta"); // See ValintatulosWithVastaanotto.persistValintatulokset check
+                    voidaanVaihtaa = false;
+                }
+            } else if (tila == ValintatuloksenTila.EI_VASTAANOTETTU_MAARA_AIKANA) {
+                TilojenMuokkaus.asetaTilaksiPerunutEiVastaanottanutMaaraaikana(hakemusWrapper);
+            } else if (tila == ValintatuloksenTila.PERUUTETTU) {
+                TilojenMuokkaus.asetaTilaksiPeruutettu(hakemusWrapper);
+            } else if (tila == ValintatuloksenTila.KESKEN) {
+                // tila == KESKEN
+                if (valintatulos.getJulkaistavissa() && hakemus.getEdellinenTila() == HakemuksenTila.HYVAKSYTTY) {
+                    hyvaksy(hakemusWrapper, valintatulos);
+                } else if (valintatulos.getJulkaistavissa() && hakemus.getEdellinenTila() == HakemuksenTila.VARASIJALTA_HYVAKSYTTY) {
+                    hyvaksyVarasijalta(hakemusWrapper, valintatulos);
+                } else if (valintatulos.getHyvaksyttyVarasijalta()) {
+                    if (hasHigherJulkaistuHyvaksytty(hakukohdeWrapper, hakemusWrapper)) {
+                        voidaanVaihtaa = true;
+                        logDidNotHyvaksy(hakemusWrapper, tila, "hyvaksyttyVarasijalta");
+                    } else {
+                        hyvaksyVarasijalta(hakemusWrapper, valintatulos);
+                    }
+                } else if (valintatulos.getHyvaksyPeruuntunut()) {
+                    if (hasHigherJulkaistuHyvaksytty(hakukohdeWrapper, hakemusWrapper)) {
+                        voidaanVaihtaa = true;
+                        logDidNotHyvaksy(hakemusWrapper, tila, "hyvaksyPeruuntunut");
+                    } else {
+                        hyvaksy(hakemusWrapper, valintatulos);
+                        hakemusWrapper.hyvaksyPeruuntunut();
+                    }
+                } else if (HakemuksenTila.HYLATTY == hakemus.getTila()) {
+                    voidaanVaihtaa = false;
+                } else {
+                    voidaanVaihtaa = true;
+                }
+            } else {
+                throw new IllegalStateException(String.format("Valintatulos %s ja hakemus %s olivat tuntemattomassa tilassa", valintatulos, hakemus));
+            }
+            hakemusWrapper.setTilaVoidaanVaihtaa(voidaanVaihtaa);
+            hakemusWrapper.getHenkilo().getValintatulos().add(valintatulos);
+        } else if (hakemus.getTila().equals(HakemuksenTila.HYLATTY)) {
+            hakemusWrapper.setTilaVoidaanVaihtaa(false);
+        } else if (aiempiVastaanottoSamalleKaudelle.isPresent()) {
+            LOG.warn("Ei muutettu hakemuksen tilaa aiemman saman kauden vastaanoton perusteella. " +
+                    "Hakukohde: {}, valintatapajono: {}, hakemus: {}, hakemuksen tila: {}, " +
+                    "hakemuksen edellinen tila: {}, vastaanoton tila: {}, aiemmin vastaanotettu hakukohde: {} {},",
+                    hakemusWrapper.getValintatapajono().getHakukohdeWrapper().getHakukohde().getOid(),
+                    hakemusWrapper.getValintatapajono().getValintatapajono().getOid(),
+                    hakemus.getHakemusOid(),
+                    hakemus.getTila(),
+                    hakemus.getEdellinenTila(),
+                    valintatulos != null ? valintatulos.getTila() : "- (ei valintatulosta)",
+                    aiempiVastaanottoSamalleKaudelle.get().getAction(),
+                    aiempiVastaanottoSamalleKaudelle.get().getHakukohdeOid());
+        }
+    }
+
+    private static void logDidNotHyvaksy(HakemusWrapper hakemusWrapper, ValintatuloksenTila tila, String flagName) {
+        Hakemus hakemus = hakemusWrapper.getHakemus();
+        LOG.info("Ei hyväksytä jonosta vaikka {} päällä, koska korkeamman prioriteetin jonossa on jo julkaistu hyväksytty: " +
+                        "Hakukohde: {}, valintatapajono: {}, hakemus: {}, hakemuksen tila: {}, " +
+                        "hakemuksen edellinen tila: {}, vastaanoton tila: {}",
+                flagName,
+                hakemusWrapper.getValintatapajono().getHakukohdeWrapper().getHakukohde().getOid(),
+                hakemusWrapper.getValintatapajono().getValintatapajono().getOid(),
+                hakemus.getHakemusOid(),
+                hakemus.getTila(),
+                hakemus.getEdellinenTila(),
+                tila);
+    }
+
+    private static boolean hasHigherJulkaistuHyvaksytty(HakukohdeWrapper hakukohdeWrapper,
+                                                        HakemusWrapper hakemusWrapper) {
+        Valintatapajono currentJono = hakemusWrapper.getValintatapajono().getValintatapajono();
+        for (ValintatapajonoWrapper toinenJono : hakukohdeWrapper.getValintatapajonot()) {
+            if (toinenJono.getValintatapajono().getPrioriteetti() < currentJono.getPrioriteetti()) {
+                for (HakemusWrapper hakemus : toinenJono.getHakemukset()) {
+                    if (hakemus.getHakemus().getHakemusOid().equals(hakemusWrapper.getHakemus().getHakemusOid())) {
+                        if (hakemus.getValintatulos().map(v -> v.getJulkaistavissa()).orElse(false)
+                                && TilaTaulukot.kuuluuHyvaksyttyihinTiloihin(hakemus.getHakemus().getEdellinenTila())) {
+                            return true;
+                        }
+                        break;
+                    }
+                }
+            }
+        }
+        return false;
+    }
+
+    private static boolean estaaVastaanotonYhdenPaikanSaannoksenTakia(Optional<VastaanottoDTO> aiempiVastaanottoOptional, HakemusWrapper hakemusWrapper) {
+        if (!aiempiVastaanottoOptional.isPresent() || hakemusWrapper.getHakukohdeOid().equals(aiempiVastaanottoOptional.get().getHakukohdeOid())) {
+            return false;
+        }
+        VastaanottoDTO aiempiVastaanotto = aiempiVastaanottoOptional.get();
+        if (aiempiVastaanotto.getAction().sitova) {
+            return true;
+        }
+        return onEriHaun(aiempiVastaanotto, hakemusWrapper);
+    }
+
+    private static boolean onEriHaun(VastaanottoDTO aiempiVastaanotto, HakemusWrapper hakemusWrapper) {
+        return hakemusWrapper.getHenkilo().getHakemukset().stream().noneMatch(h ->
+            h.getHakukohdeOid().equals(aiempiVastaanotto.getHakukohdeOid()));
+    }
+
+    private static void hyvaksyVarasijalta(final HakemusWrapper hakemusWrapper, final Valintatulos valintatulos) {
+        TilojenMuokkaus.asetaTilaksiVarasijaltaHyvaksytty(hakemusWrapper);
+        hakemusWrapper.getHakemus().setIlmoittautumisTila(valintatulos.getIlmoittautumisTila());
+    }
+
+    private static void hyvaksy(final HakemusWrapper hakemusWrapper, final Valintatulos valintatulos) {
+        TilojenMuokkaus.asetaTilaksiHyvaksytty(hakemusWrapper);
+        hakemusWrapper.getHakemus().setIlmoittautumisTila(valintatulos.getIlmoittautumisTila());
+    }
+
+    private static boolean viimeisinHyvaksyttyJono(HakemusWrapper hakemusWrapper) {
+
+        if(TilaTaulukot.kuuluuHyvaksyttyihinTiloihin(hakemusWrapper.getHakemus().getEdellinenTila())) return true;
+
+        final String viimeisinHyvaksyttyJonoOid = hakemusWrapper.getHenkilo().getHakemukset().stream()
+                .filter(hw -> hw.getHakemus().getTilaHistoria().stream().anyMatch(th -> TilaTaulukot.kuuluuHyvaksyttyihinTiloihin(th.getTila())) == true)
+                .map(hw -> {
+                    final TilaHistoria tilaHistoria = hw.getHakemus().getTilaHistoria().stream()
+                            .filter(th -> TilaTaulukot.kuuluuHyvaksyttyihinTiloihin(th.getTila()))
+                            .max(Comparator.comparingLong(th -> th.getLuotu().getTime()))
+                            .get();
+                    org.apache.commons.lang3.tuple.Pair<String, Date> pair = Pair.of(hw.getValintatapajono().getValintatapajono().getOid(), tilaHistoria.getLuotu());
+
+                    return pair;
+
+                }).collect(Collectors.toList()).stream()
+                .max(Comparator.comparingLong(pair -> pair.getRight().getTime())).orElse(Pair.of("", null)).getLeft();
+
+        return viimeisinHyvaksyttyJonoOid.equals(hakemusWrapper.getValintatapajono().getValintatapajono().getOid());
+
+
     }
 }
